WAS-main/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.16
botocore==1.34.16
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
distro==1.9.0
Django==5.0.1
django-allauth==0.60.0
django-ckeditor==6.7.0
django-countries==7.5.1
django-environ==0.11.2
django-filter==23.5
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-pragmatic==4.8.0
django-storages==1.14.2
django-whisper==0.5.0
exceptiongroup==1.2.0
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.7.1
pgvector==0.2.4
phonenumbers==8.13.27
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-pragmatic==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
zope.interface==6.1
